some changes

+++ snowman.hpp
@@ -2,5 +2,5 @@
 using namespace std;
 
 namespace ariel{
-	string snowman(long num);
+	string snowman(int num);
 };
